Added persist-model feature
Now when creating a RecommendationEngine object, we also have
to provide the model_path to where the model will be persisted/saved.
By default, after every time a new model is created
(at the first running, after new user ratings are sent to the model),
the model is saved in the model_path.
When starting the app, if a previously saved model is present
at model_path, he will be loaded by default.
To force the engine to train a new model, just
delete the saved_models/movie_lens_als/ directory and
start the app again.
$ sudo rm -r saved_models/movie_lens_als/