Added facebook login button, gets response data
Properly configured webpack and npm packages.
Use "$ npm install" to install the required dependencies,
"$ npm run-script build" to build all the npm packages and
React components in app/static/bundle.js file, that is
added as a script file in the app/templates/client/index.html.
At this point all the frontend is generated in bundle.js, just
run the spark-submit command as usual to run the server on
Spark and access http://0.0.0.0:5435.
More in README.md.